views, urls updated
